# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TÖPP API",
    "description": "REST API интернет-магазина ковров; цены в целых рублях. /docs для интерактивного OpenAPI.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "summary": "List Products",
        "operationId": "list_products_api_v1_products_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "style",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Style"
            }
          },
          {
            "name": "colorCode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Colorcode"
            }
          },
          {
            "name": "countryCode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Countrycode"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Size"
            }
          },
          {
            "name": "lengthCm",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lengthcm"
            }
          },
          {
            "name": "widthCm",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Widthcm"
            }
          },
          {
            "name": "minPrice",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minprice"
            }
          },
          {
            "name": "maxPrice",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Maxprice"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(popular|newest|price_asc|price_desc|size|discount)$",
              "default": "newest",
              "title": "Sort"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog/filters": {
      "get": {
        "summary": "Filters",
        "operationId": "filters_api_v1_catalog_filters_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{slug}": {
      "get": {
        "summary": "Product Detail",
        "operationId": "product_detail_api_v1_products__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "cartBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Токен из POST /api/v1/carts"
      },
      "adminBasic": {
        "type": "http",
        "scheme": "basic",
        "description": "ADMIN_USERNAME и пароль администратора"
      }
    }
  }
}